+++ CHANGELOG.md
@@ -28,7 +28,7 @@
   [Full Changelog](https://github.com/pubnub/javascript/compare/v4.13.0...v4.14.0)
 
 
-- ⭐Allow disable of heartbeats by passing interval = 0
+- ⭐Allow disable of heartbeats by passing heartbeatInterval = 0